requirements.txt
google-generativeai
uvicorn
pydantic
fastapi
Pillow
python-dotenv
requests